删除eip前,先对eip是否存在进行判断.(感觉还是hibernate关联策略的问题)

+++ src/main/java/com/sobey/cmop/mvc/service/iaas/EipService.java
@@ -87,7 +87,9 @@ public class EipService extends BaseSevcie {
 
 	@Transactional(readOnly = false)
 	public void deleteNetworkEipItem(Integer id) {
-		networkEipItemDao.delete(id);
+		if (this.getNetworkEipItem(id) != null) {
+			networkEipItemDao.delete(id);
+		}
 	}
 
 	/**
